Remove `simplession.Session` dependency from `Store` interface{} to /v2.
This is a breaking change. It removes the superfluous dependency
stores had on the simplesessions package solely for accessing string
generation functions, which are not needed in sessions. ID generation
is the purview of individual store implementations.

It also adds a go.mod to the `conv` package as its referenced by
several stores.