Refactor all stores to remove the simplesessions dependency and adhere to the new `simplesessions/v2` interface methods. This is a breaking change.`
This commit also refactors securestore and changes its behaviour.

This store is peculiar because it does not actually have a persistent
backend store and instead uses the session/cookie ID string (value)
itself as the encoded store. The older version thus relied on Session{}
to write the cookie to the frontend, which is not ideal.

This version removes the dependency and leaves the actual cookie writing
to the implementer. This commit introduces a new `Flush()` method that
returns the encoded values set using Set(), which the implementer can
then write to a cookie externally (with a `Session.WriteCookie()`).